@@ -4,20 +4,28 @@
   "description": "Search ECMAScript Structurally",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "grammar": "node ./node_modules/pegjs/bin/pegjs grammar.pegjs parser.js"
+    "prebuild": "mkdirp lib & npm run build-grammar",
+    "build": "babel src -d lib",
+    "build-grammar": "pegjs -o lib/queryParser.js src/query-grammar.pegjs"
   },
   "keywords": [],
   "author": "James Donaghue",
   "license": "ISC",
   "engines": {
-    "node": ">=6.0"
+    "node": ">=10.12.0"
   },
   "dependencies": {
-    "acorn": "^3.0.4",
-    "colors": "^1.1.2"
+    "acorn": "^6.0.4",
+    "acorn-jsx": "^5.0.0",
+    "acorn-loose": "^6.0.0",
+    "acorn-walk": "^6.1.0",
+    "colors": "^1.3.1"
   },
   "devDependencies": {
-    "pegjs": "^0.9.0"
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.5",
+    "@babel/preset-env": "^7.1.5",
+    "mkdirp": "^0.5.1",
+    "pegjs": "^0.10.0"
   }
 }
